requirements.txt
astroid==2.3.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
idna==2.8
importlib-metadata==0.23
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pkg-resources==0.0.0
pluggy==0.13.0
py==1.8.0
pylint==2.4.1
pyparsing==2.4.2
pytest==5.2.0
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
